The app is a game rental that allows users to rent gaming consoles, controllers, and game discs.
Application Details
● Login/Register Functionality:
○ The user should be able to register as a Seller or a
Gamer. Please note a user can either be a Seller or
Gamer, he/she cannot be both.
○ The users should be able to log in using Username and
Password
○ Post login if the user is a seller then show the following
menu:
■ Create Product
■ Update Product
■ View all Products
■ View Product Details by ID
■ View User Profile - A seller can see only the
profiles of the gamers and not other sellers.
■ Logout
○ Post login if the user is a gamer then show the following
menu:
■ View all Product
■ Manage wishlist
■ Manage cart
■ Place order
■ Update Profile
■ View Orders
■ Logout
● Create Product Functionality:
○ The seller should be able to create a product. They can
create the product using following details:
■ ProductID - Autogenerated value to uniquely
identify a product. User should not be asked to
enter this.
■ Product Title
■ Product Type - The product type be any of the
follows: game, controller, or console.
■ Price Per Day. For eg, 100, 400 or 1000.
■ Total Stock Available
○ After the Product is created the details of that product
should be stored in the Products.json file.
● Update Product Functionality:
○ The seller should be able to update a product. They can
update the following details:
■ Product Title
■ Product Type
■ Price Per Day
■ Total Stock Available
○ In case of update functionality, based on ProductID, allow the
seller to update the details.
○ After the Product is updated the details of that product should
be stored in the Products.json file.
● View All Products Functionality:
○ If the current logged-in user is a seller then the products
created by the logged-in seller should be displayed to
him/her and not from all the sellers.
○ If the current logged-in user is a gamer then show the
list of all the products from all the sellers.
○ For each product following details need to be displayed
■ ProductID
■ Product Title
■ Product Type
■ Price Per Day
■ Total Stock Available
● Product Details Functionality:
○ Based on the product ID show only the details of that product.
● View User Profile:
○ Based on the username show all the details of the gamer.
● Manage Wishlist:
○ When a gamer selects this then show two options:
■ Add Item to Wishlist: Request user to enter productID of
the product and then add it in the logged-in gamer’s
wishlist.
■ Remove items from Wishlist: Request user to enter
productID of the product and then remove it from the
logged-in gamer’s wishlist.
○ Wishlist data will be stored in the Gamers.json file with the
gamer data.
● Manage Cart:
○ When a gamer selects this then show three options:
■ Add Item to Cart: Request user to enter the following
product details and then add it in the logged-in gamer’s
cart - ProductID, Quantity(to be purchased), Booking
Start Date(in format YYYY-MM-DD), and Booking End
Date(in format YYYY-MM-DD). Multiple items can be
added to cart by choosing this option multiple times.
■ Remove items from Cart: Request user to enter productID
of the product and then remove it from the logged-in
gamer’s cart.
■ View Cart: Show all the products in the gamer’s cart.
○ Cart data will be stored in the Gamers.json file with the gamer
data.
● Place Order Functionality:
○ When the gamer selects this option then simply place the
order for all the products in his/her cart.
○ The order total price will be calculated based on the per day
price of the products and the number of days they have been
rented.
○ Order data will be stored in the orders.json file.
● View Orders:
○ Show the list of all the orders for the current logged-in gamer.
● Update Profile:
○ Give gamer an option to update - Email, Password, or Contact
Number.
Common Functionalities:
● If any of the validations fail then show appropriate error
messages.
● Validations:
○ Length of password > 4
○ The usernames for gamer and sellers should be unique.
○ Length of contact number. = 10
○ Seller’s email should be ending with @admin.com. For
example, qaifi@admin.com
○ No empty fields
● You have to store all the data in the given JSON files.
What is already implemented in the base project?
● Registration functionality for both seller and gamer is
implemented
● Login Functionality for both seller and gamer is implemented.
● The interactive command line for application users is already
implemented.
● The code to take inputs from users is already implemented in the
commandline.py file. You DO NOT have to write extra code to take
inputs from the users.
What do you have to implement in the project for submission?
● Implement all the remaining functionalities in the operations.py
file.
● Please ensure that you update the data in the respective JSON
files. The grading team will verify the JSON files. Even if your
functionality is working if you haven’t updated the JSON file
correctly then also you will receive 0 marks for that specific
functionality.
● All the seller data will be managed in the Sellers.json file.
● All the gamer data will be managed in the Gamers.json file.
● All the product data will be managed in the Products.json file.
● All the order data will be managed in the Orders.json file.
Instructions for Code Execution:
● Running the application in Terminal/Command Prompt:
○ Download the following code: CODE FILE LINK
○ Extract the code from the zip file.
○ Open the project code folder in VS Code.
○ Open Terminal in the VS Code
○ Run commandline.py file by executing the following
command “python commandline.py”
● Please note, every time there is any change in the code, re-run
the command -> “python commandline.py”. If it is already
running then press Ctrl + C to exit it and run the above command
again. This will help you to find the errors in your code if any.
Submission Details
● You have to submit the code URL. Please upload all your code in a
folder on Google drive. Please ensure that the folder is available
publicly or else the grading team will give you zero marks. To check
that please try to open the URL in incognito mode. You have to share
the public URL of this google drive folder.
● Please ensure that you DO NOT COPY the code from your batch
mates or the internet. Our proctoring tool will detect plagiarism
and you will receive zero marks in the exam.
● Please DO NOT share your code publicly. If anyone copies your
code then both you and the other person will receive zero marks.
● You DO NOT have to submit the project video.
● You have to submit the code URL on the certification details page on
the classroom portal. Follow the steps below:
○ Goto Round 2 card on the certification details page
○ Click on the “Submit Project” button
○ Fill the form in the popup and submit. Your project does not
require a hosted web app or Video URL so you can simply
paste the code URL itself in all 3 fields.
● Please ensure that you make your submissions before the due date
and time. No late submissions will be accepted